fix osx

+++ test/integration/integration.cc
@@ -57,7 +57,8 @@ void IntegrationStreamDecoder::waitForHeaders() {
 void IntegrationStreamDecoder::waitForBodyData(uint64_t size) {
   ASSERT(body_data_waiting_length_ == 0);
   body_data_waiting_length_ = size;
-  body_data_waiting_length_ -= std::min(body_data_waiting_length_, body_.size());
+  body_data_waiting_length_ -=
+      std::min(body_data_waiting_length_, static_cast<uint64_t>(body_.size()));
   if (body_data_waiting_length_ > 0) {
     dispatcher_.run(Event::Dispatcher::RunType::Block);
   }
